requirements.txt
streamlit
fastapi
uvicorn
openai
tiktoken
langchain
langchain-core
langchain-community
langchain-chroma
langchain-pinecone
chromadb
pinecone-client
pypdf
python-dotenv
requests
openai